Remove references to upstream repositories and files and update README
## README

* Adds more details about our project and relationship to the original Asterinas project.
* Fixes incorrect docker image tag

## Upstream References

This removes all references to:

* Asterinas github org repositories, replacing them with repositories in this org.
* Asterinas dockerhub images, replacing them with references to https://hub.docker.com/u/ldosproject

This also performs some changes to avoid having to copy quite as much infrastructure into our own organizations:

* The Linux kernel for benchmarking is built as part of the docker image build (along with all the benchmark apps themselves). Previously, this was downloaded from an Asterinas controlled "cache" repo.
* The vdso is extracted from an ubuntu package (matching the build environment as with the rest of the Asterinas userspace). Previously, this was downloaded like the kernel.
* Update the inventory crate which avoids the need for a patched dependency since it no longer uses the ctor crate. This eliminates the need for customized versions of both inventory and ctor.

## Github Actions

These changes make no attempt to leave the github actions in a working state; they simply remove upstream references. The actions are probably somewhat broken due to infrastructure expected by the actions is not available in the LDOS organization and repos. A future PR will have to reenable github actions. This entirely removes some of the support for TDX (Intel® Trust Domain Extensions) in the github actions testing configuration since it will not be needed for our work.