Fix KVM device check in makefile. (#51)
A change in aea2f3f made tests run without KVM by default. This exposed a bug in the microvm boot: #50.

This PR fixes the makefile to use KVM by default so that the microvm CI runs will succeed. This PR also disable stripping of the elf for microvm builds to simplify debugging of any future failures.

This fixes the microvm panic issue discussed in #22, but not the test failures that do not panic.