Remove UnwindSafe bound on ORPC servers. (#84)
This removes the requirement that server state implement UnwindSafe.

## Justification

Server state is already required to be `Sync`, meaning that it has to guarantee the safety of concurrent accesses. Observing the state of data during or after a panic is no different.

When any part of a server panics, all parts of the server should panic. This is a form of server-wide poisoning where all execution in the server is stopped.

Synchronization primitives like `Mutex` can include checks which immediately trigger a panic if the server as a whole has panicked. This would prevent observing the poisoned internals of a Mutex (with some careful ordering of when to set the "panicked flag"). Note: using a panic here instead of `Result` is reasonable because we isolate servers from each others panics.

UnwindSafe is mostly advisory anyway, as it is trying to provide some degree of semantic safety, but Rust in general does not guarantee that. In `std`, it is possible to access the poisoned data (see `PoisonError`). We expect to support this too, so that servers can be restarted without losing their entire state.