Add "weak this" support to ORPC. (#86)
This adds support for servers that are cyclic in that they contain a weak reference to themselves. See Arc::new_cyclic. This is a small extension because the weak this was already used internally.

NOTE: The "weak this" pattern is arguably an anti-pattern, but it is common and needs to be supported.