Split HugePages when required (#136)
Previously if a region that was mapping as huge by hugepaged was remapped, the remap could cause a kernel panic if the new address wasn't aligned to the huge page size. This PR changes the behavior of remapping to first break up the huge page.

Additionally, HugePage splitting was not correctly updating the frame to track the new level and reference count.